Issue #6
Modified the excel file.  Converted excel file to gdb table.  Converted gdb table to point feature class.  Converted point feature class to json.  Also Including in the commit an mxd we can utilize to manage the data if desired.

Only issue I had is that the json is a 166 MB file.  Github didn't like that.  I'll host it for us here instead and we can just link our web app to the json directly here on my website:

http://rkpalmerjr.com/GEOG575_Finalproject/Florida_NAS.json